+++ Core/Interfaces/IGrupoBusiness.cs
@@ -4,9 +4,9 @@ public interface IGrupoBusiness
 {
     public Task<Response<GrupoDTO>> CreateGrupoAsync(GrupoInsertDTO insertDTO);
 
-    public Task<Response<IEnumerable<UsuarioDTO>>> GetUsuariosAsync(int idGrupo);
+    public Task<Response<IEnumerable<UsuarioDTO>>> GetUsuariosAsync(long idGrupo);
 
-    public Task<Response<IEnumerable<RecordatorioDTO>>> GetRecordatoriosAsync(int idGrupo);
+    public Task<Response<IEnumerable<RecordatorioDTO>>> GetRecordatoriosAsync(long idGrupo);
 
     public Task<Response<GrupoWithUserDTO>> InsertToGrupoAsync(int idCodigo, long idUsuario);
