+++ Core/Interfaces/IRecordatorioBusiness.cs
@@ -4,11 +4,11 @@ namespace NotasAPI.Core.Interfaces;
 
 public interface IRecordatorioBusiness
 {
-    Task<Response<RecordatorioInsertDTO>> CreateRecordatorio(int idUsuario, RecordatorioInsertDTO insertDTO);
+    Task<Response<RecordatorioInsertDTO>> CreateRecordatorio(long idUsuario, RecordatorioInsertDTO insertDTO);
 
-    Task<Response<IEnumerable<RecordatorioDTO>>> GetAllRecordatorios(int idUsuario);
+    Task<Response<IEnumerable<RecordatorioDTO>>> GetAllRecordatorios(long idUsuario);
 
-    Task<Response<RecordatorioWithGroupDTO>> CreateRecordatorioInGroup(int idUsuario, int idGrupo, RecordatorioInsertDTO insertDTO);
+    Task<Response<RecordatorioWithGroupDTO>> CreateRecordatorioInGroup(long idUsuario, long idGrupo, RecordatorioInsertDTO insertDTO);
 
     Task<Response<RecordatorioUpdateDTO>> UpdateRecordatorio(RecordatorioUpdateDTO updateDTO);
