+++ Repositories/Interfaces/IGrupoRepository.cs
@@ -4,9 +4,9 @@ public interface IGrupoRepository : IRepository<Grupo>
 {
     public Task<GrupoDTO> CreateGrupoAsync(GrupoInsertDTO insertDTO, int codigo);
 
-    public Task<IEnumerable<UsuarioDTO>> GetUsuariosAsync(int idGrupo);
+    public Task<IEnumerable<UsuarioDTO>> GetUsuariosAsync(long idGrupo);
 
-    public Task<IEnumerable<RecordatorioDTO>> GetRecordatoriosAsync(int idGrupo);
+    public Task<IEnumerable<RecordatorioDTO>> GetRecordatoriosAsync(long idGrupo);
 
     public Task<GrupoWithUserDTO> InsertToGrupoAsync(GrupoDTO grupo, Usuario usuario);
